Allow courses to stop registration
The enrollment button is currently active while a course is running,
allowing random strangers to "enroll" when registration has already
ended.
This change should allow a course to set an end date for registrations,
limiting further attendance changes to admins.

This is added to the online course, which suffer from the problem.